backend/requirements.txt
Flask==3.1.2
Flask-CORS==5.0.0
opencv-python==4.10.0.84
cvzone==1.6.1
mediapipe==0.10.20
requests==2.31.0
numpy>=1.21.0
protobuf>=3.20.0
# Email dependencies
Flask-Mail==0.9.1
email-validator==2.1.0
python-dotenv==1.0.0